No need for FS outside of interop, for now
